Implementing BasicLotterySampling for the Heavy Hitters problem. Refactoring + cleaning code

+++ src/algorithms/lottery_sampling/Algorithm.h
@@ -32,7 +32,7 @@ public:
 
     FrequencyOrder<Element<T>>& get_frequency_order() override;
 
-    float get_threshold() const override;
+    double get_threshold() const override;
 
     void print_state() override;
 };
